+++ ui/src/components/dialog/SettingsDialog.vue
@@ -1,47 +1,23 @@
 <template>
-  <Dialog :title="t('settingsTitle')" :show="show" @close="close" :width="`700px`" :goBack="close">
+  <Dialog :title="t('settingsTitle')" :show="show" @close="close" width="700px" :goBack="close" hide-actions>
     <template v-slot:content>
-      <v-card-text class="pt-4">
-        <h3>{{ t('keyBackupTitle') }}</h3>
-        <p>
-          {{ t('loseAccess') }}.
-          {{ t('backUpKey') }}
-        </p>
-        <v-col style="margin-bottom:20px" align="left" class="py-0 pl-0">
-          <v-btn
-            class="white--text"
-            color="black"
-            @click.prevent="downloadKeys"
-          >
-            {{ t('downloadKey') }}
-          </v-btn>
-      </v-col>
-        <h3 class="pt-6">{{ t('languageTitle') }}</h3>
-        <p>
-          {{ t('chooseLanguage') }}
-        </p>
-        <LocalePicker class='locale' />
-        <p>
-          {{ t('languagesNote') }}
-          <a href="http://chat.ahau.io" target="_blank">{{ t('submitFeedback') }}</a>
-        </p>
-      </v-card-text>
-    </template>
-    <template v-slot:actions>
-      <v-col align="center" class="py-0">
-      </v-col>
+      <v-container fluid>
+        <LanguageSettings />
+        <v-divider />
+        <BackupSettings />
+        <v-divider />
+        <StorageSettings />
+      </v-container>
     </template>
   </Dialog>
 </template>
 
 <script>
 import Dialog from '@/components/dialog/Dialog.vue'
-import LocalePicker from '@/components/LocalePicker'
 
-import { createNamespacedHelpers } from 'vuex'
-import { downloadBackup } from '@/lib/download-helper'
-
-const { mapActions: mapSettingsActions } = createNamespacedHelpers('settings')
+import LanguageSettings from '@/components/settings/LanguageSettings.vue'
+import StorageSettings from '@/components/settings/StorageSettings.vue'
+import BackupSettings from '@/components/settings/BackupSettings.vue'
 
 export default {
   props: {
@@ -50,17 +26,11 @@ export default {
   name: 'SettingsDialog',
   components: {
     Dialog,
-    LocalePicker
+    BackupSettings,
+    LanguageSettings,
+    StorageSettings
   },
   methods: {
-    ...mapSettingsActions(['getBackup', 'updateKeyBackupSettings']),
-    async downloadKeys () {
-      const backupContent = await this.getBackup()
-
-      downloadBackup(backupContent)
-
-      await this.updateKeyBackupSettings(true)
-    },
     close () {
       this.$emit('close')
     },
@@ -70,8 +40,11 @@ export default {
   }
 }
 </script>
+
 <style scoped lang="scss">
-.locale {
-    margin-bottom:20px;
-  }
+
+// v-divider
+hr {
+  margin-bottom: 20px;
+}
 </style>